+++ packages/flutter_riverpod/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased build
+
+- Added support for `Ref/ProviderContainer.invalidate(provider, asReload: true)`
+
 ## 3.0.0-dev.3 - 2023-11-27
 
 - Fix "pending timer" issue inside tests when using `ref.keepAlive()`.