+++ packages/riverpod/CHANGELOG.md
@@ -4,6 +4,7 @@
 - **Breaking** removed `AsyncValue.valueOrNull` (use `.value` instead).
 - `Stream/FutureProvider.overrideWithValue` was added back.
 - Updated `AsyncValue` documentations to use pattern matching.
+- Added support for `Ref/ProviderContainer.invalidate(provider, asReload: true)`
 
 ## 3.0.0-dev.3 - 2023-11-27